server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -7,9 +7,10 @@ const NotFoundError = require('../../exceptions/NotFoundError');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class NotesService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
+    this._cacheService = cacheService;
   }
 
   async verifyNoteOwner(noteId, userId) {
@@ -63,20 +64,32 @@ class NotesService {
       throw new InvariantError('Catatan gagal ditambahkan');
     }
 
+    await this._cacheService.delete(`notes:${owner}`);
     return rows[0].id;
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes 
+    try {
+      const result = await this._cacheService.get(`notes:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes 
       LEFT JOIN collaborations ON collaborations.note_id = notes.id 
       WHERE notes.owner=$1 OR collaborations.user_id=$1 
       GROUP BY notes.id`,
-      values: [owner],
-    };
+        values: [owner],
+      };
 
-    const { rows } = await this._pool.query(query);
-    return rows.map(mapDBToModel);
+      const { rows } = await this._pool.query(query);
+      const mappedResult = rows.map(mapDBToModel);
+
+      await this._cacheService.set(
+        `notes:${owner}`,
+        JSON.stringify(mappedResult)
+      );
+      return mappedResult;
+    }
   }
 
   async getNoteById(id) {
@@ -100,30 +113,36 @@ class NotesService {
   async editNoteById(id, { title, body, tags }) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id=$5 RETURNING id',
+      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id=$5 RETURNING id,owner',
       values: [title, body, tags, updatedAt, id],
     };
 
-    const { rowCount } = await this._pool.query(query);
+    const { rowCount, rows } = await this._pool.query(query);
 
     if (!rowCount) {
       throw new NotFoundError(
         'Gagal memperbaharui catatan. Id tidak ditemukan'
       );
     }
+
+    const { owner } = rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id=$1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id=$1 RETURNING id,owner',
       values: [id],
     };
 
-    const { rowCount } = await this._pool.query(query);
+    const { rowCount, rows } = await this._pool.query(query);
 
     if (!rowCount) {
       throw new NotFoundError('Gagal menghapus catatan. Id tidak ditemukan');
     }
+
+    const { owner } = rows[0];
+    await this._cacheService.delete(owner);
   }
 }
